Crate icu_locale_fallback

Tools for locale fallback, enabling arbitrary input locales to be mapped into the nearest locale with data.

Modules

provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

Structs

LocaleFallbackConfig

Configuration settings for a particular fallback operation.

LocaleFallbackIterator

Iteration type for locale fallback operations.

LocaleFallbacker

Implements the algorithm defined in UTS #35: Locale Inheritance and Matching.

LocaleFallbackerBorrowed

Borrowed version of LocaleFallbacker.

LocaleFallbackerWithConfig

A LocaleFallbackerBorrowed with an associated LocaleFallbackConfig.

Enums

LocaleFallbackPriority

Hint for which subtag to prioritize during fallback.