icu_locale_fallback

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

Baked

Baked data

LikelySubtagsForLanguage

This likely subtags data is used for the minimize and maximize operations.

LocaleLikelySubtagsLanguageV1

Marker for data for likely subtags for languages.

LocaleParentsV1

Marker for locale fallback parents data.

Parents

Locale fallback rules derived from CLDR parent locales data.